made dashboard handler
