+++ src/scripts/handlers/welcomeHandler.js
@@ -1,8 +1,11 @@
+// console.log("welcome handler loaded")
+
 import { Act } from "../actions.js";
 import { API } from "../api.js";
+import { dom } from "../dom.js";
 
 export const welcomeHandler = {
-  login(showDashboard) {
+  login() {
     Act.click("#login-button", () => {
       let username = document.querySelector("#username-input").value
       if (!username) alert("You must enter a username")
@@ -14,9 +17,9 @@ export const welcomeHandler = {
             })
             if (!validUser[0]) alert("invalid username")
             else {
-              console.log(`Hi ${validUser[0].name}, your id is ${validUser[0].id}`)
+              console.log(`${validUser[0].name}, id = ${validUser[0].id}`)
               sessionStorage.setItem("activeUser", validUser[0].id)
-              showDashboard()
+              dom.displayDashboard()
             }
         })
       }
